requirements.txt
openai
streamlit
python-dotenv
google-generativeai
anthropic
spacy
chromadb
pypdf
langchain-openai
langchain-community
huggingface_hub
beautifulsoup4
langchain
langchain-ollama
langchain-chroma
langchain-google-genai 
pandas
PyPDF2
datasets[audio]
transformers
torch
tensorflow
ipywidgets
datasets
accelerate
evaluate
timm
